Correct useless bug in day 1
Realized reading through the solution thread that there was an error in
my solution, it just wasn't hit by the input given. I was
double-counting numbers, which would have been a problem if 1010 was in
the input.

I switched to using `_.combinations()` to ensure I'm using unique
combinations, but it's back to being slow on large inputs since it's no
longer doing the early filtering it was before.